Major changes so that deltaTime is taken into account
- changes so that everything is updating based on deltaTime (including
  rate of fire, fuel, movement, and lifetime)
- added deltaTime to Game class so collisions could access it
- fixed collision handling so that it takes deltaTime into consideration
- removed indicator class as it was redundant with ObjectiveIndicator
  class
- added lifetime variable to missiles
- updated TODO
- added documentedBugs.txt as a place to list defects separate from TODO

+++ Turret.cpp
@@ -3,7 +3,7 @@
 #include "Team.h"
 #include "Player.h"
 
-Turret::Turret(unsigned int index, Player* o, renderableType i, Base* follow, Vec2D loc, Vec2D dir, unsigned int Rof, bool collides, bool draw) : Weapon(index, o, follow, loc, dir, Rof, collides, draw)
+Turret::Turret(unsigned int index, Player* o, renderableType i, Base* follow, Vec2D loc, Vec2D dir, float Rof, bool collides, bool draw) : Weapon(index, o, follow, loc, dir, Rof, collides, draw)
 {
 	mesh = CopyMesh(i, (o? getGameType()->teams[o->team].color:Color()));	//each object's mesh will have an individual color, so each has its own copy
 	if(collidable)
