+++ Turret.h
@@ -6,7 +6,7 @@
 class Turret : public Weapon
 {
 public:
-	Turret(unsigned int index, Player* o, renderableType mesh, Base* follow, Vec2D loc, Vec2D dir, unsigned int Rof, bool collides = false, bool draw = false);
+	Turret(unsigned int index, Player* o, renderableType mesh, Base* follow, Vec2D loc, Vec2D dir, float Rof, bool collides = false, bool draw = false);
 	virtual ~Turret(void);
 
 	virtual void shoot();
